New project structure
Wrap the API in a namespace and adopt a new file name convention.
